+++ include/calculator/motor/EstimateFLA.h
@@ -27,7 +27,7 @@ public:
      */
     EstimateFLA(
         double motorRatedPower,
-        int motorRPM,
+        double motorRPM,
         Motor::LineFrequency lineFrequency,
         Motor::EfficiencyClass efficiencyClass,
         double specifiedEfficiency,
@@ -61,7 +61,7 @@ private:
     /// Rated Power of motor
     double motorRatedPower_ = 0.0;
     /// Motor RPM
-    int motorRPM_ = 0;
+    double motorRPM_ = 0;
     /// Line Frequency of Motor
     Motor::LineFrequency lineFrequency_;
     /// Efficiency class of Motor
